requirements.txt
altair==5.1.1
annotated-types==0.5.0
appnope==0.1.3
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
blis==0.7.11
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudpathlib==0.15.1
comm==0.1.4
confection==0.1.3
cymem==2.0.8
debugpy==1.8.0
decorator==5.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
executing==2.0.0
gitdb==4.0.10
GitPython==3.1.37
htbuilder==0.6.2
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.2
langcodes==3.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.1.0
murmurhash==1.0.10
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.26.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
preshed==3.0.9
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.6.0
rpds-py==0.10.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
spacy==3.7.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
st-annotated-text==4.0.1
stack-data==0.6.3
streamlit==1.27.1
tenacity==8.2.3
thinc==8.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
typer==0.7.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.5
validators==0.22.0
wasabi==1.1.2
wcwidth==0.2.8
weasel==0.3.1
zipp==3.17.0
